// data/modelPhotos.generated.ts
export interface ModelPhoto { name: string; image: string; }
export interface ModelSeries { name: string; models: ModelPhoto[]; }